version 2.8.0 with: Convert tag to an Object Tag with it's mongodb id. Convert all model Objects to mongodb documents. Simplify the parametritzed calls and restrict them only to query methods where are usefull only. Implements and prove the remove methods for tag and unity using repositories delete function.

+++ src/main/java/org/dmr/services/impl/HowDYLServiceImpl.java
@@ -1,20 +1,20 @@
 package org.dmr.services.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.dmr.domain.impl.Robot;
+import org.dmr.domain.impl.Tag;
 import org.dmr.domain.impl.UnityKnowledgeObject;
 import org.dmr.repositories.RobotRepository;
+import org.dmr.repositories.TagRepository;
 import org.dmr.repositories.UnityKnowledgeObjectRepository;
 import org.dmr.services.HowDYLService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoOperations;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
-
-import com.mongodb.WriteResult;
 
 /**
  * Created by davidmartinezros on 22/04/2017.
@@ -31,6 +31,9 @@ public class HowDYLServiceImpl implements HowDYLService {
 	
 	@Autowired
 	UnityKnowledgeObjectRepository repositoryUnity;
+	
+	@Autowired
+	TagRepository repositoryTag;
 	
 	@Autowired
 	MongoOperations mongoOperation;
@@ -60,17 +63,29 @@ public class HowDYLServiceImpl implements HowDYLService {
     }
     
     @Override
-    public Robot removeRobot(Robot robot) {
+    public Robot removeRobot(String idRobot) {
     	
-    	robotRepository.delete(robot);
+    	Robot robot = robotRepository.findById(idRobot);
+    	
+    	if(robot != null) {
+    		
+    		robotRepository.delete(robot);
+    	}
     	
     	return robot;
     }
     
     @Override
-    public Robot getRobot(String name) {
+    public Robot getRobotByName(String name) {
     
     	return robotRepository.findByName(name);
+    	
+    }
+    
+    @Override
+    public Robot getRobot(String idRobot) {
+    	
+    	return robotRepository.findById(idRobot);
     	
     }
     
@@ -97,28 +112,28 @@ public class HowDYLServiceImpl implements HowDYLService {
     }
     
     @Override
-    public UnityKnowledgeObject updateUnity(UnityKnowledgeObject unity) {
+    public UnityKnowledgeObject removeUnity(String idUnity) {
     	
-    	unity = repositoryUnity.save(unity);
+    	UnityKnowledgeObject unity = repositoryUnity.findById(idUnity);
+    	
+    	if(unity != null) {
+    		// El delete de unity no acaba d'anar be
+    		// esborra l'element pero retorna null i acaba fent 3 reintents
+    		repositoryUnity.delete(unity);
+    	
+    	}
     	
     	return unity;
     	
     }
     
     @Override
-    public void removeUnity(String keyRobot, Object valueRobot, String key, Object value) {
+    public UnityKnowledgeObject updateUnity(UnityKnowledgeObject unity) {
     	
-    	Query query = new Query();
-		query.addCriteria(
-			Criteria.where(keyRobot).is(valueRobot)
-		);
-		Update update = new Update();
-		update.pull("unities", new Query().addCriteria(Criteria.where(key).is(value)));
-		
-		WriteResult result = mongoOperation.updateFirst(query, update, Robot.class);
-		
-		System.out.println(result.toString());
-		
+    	unity = repositoryUnity.save(unity);
+    	
+    	return unity;
+    	
     }
 
     @Override
@@ -137,24 +152,23 @@ public class HowDYLServiceImpl implements HowDYLService {
     	
     	}
     	
-    	return null;
+    	return new ArrayList<UnityKnowledgeObject>();
     }
     
     @Override
-    public void deleteRelation(String keyUnity, Object valueUnity, String key, Object value) {
+    public UnityKnowledgeObject removeRelation(String idRelation) {
     	
-    	Query query = new Query();
-		query.addCriteria(
-			Criteria.where(keyUnity).is(valueUnity)
-		);
-		Update update = new Update();
-		update.pull("relations", new Query().addCriteria(Criteria.where(key).is(value)));
+    	UnityKnowledgeObject relation = repositoryUnity.findById(idRelation);
+    	
+    	if(relation != null) {
+    	
+    		repositoryUnity.delete(relation);
+    	
+    	}
+    	
+    	return relation;
 		
-		WriteResult result = mongoOperation.updateFirst(query, update, UnityKnowledgeObject.class);
-		
-		System.out.println(result.toString());
-		
-    }    
+    }
     
     @Override
     public UnityKnowledgeObject getUnity(String keyRobot, Object valueRobot, String key, Object value) {
@@ -206,6 +220,52 @@ public class HowDYLServiceImpl implements HowDYLService {
 		
 		return null;
 		
+    }
+    
+    @Override
+    public Tag createTag(UnityKnowledgeObject unity, Tag tag) {
+        
+    	repositoryTag.save(tag);
+    	
+    	unity.addTag(tag);
+    	
+    	unity = repositoryUnity.save(unity);
+    	
+        return tag;
+        
+    }
+    
+    @Override
+    public Tag removeTag(String idTag) {
+    	
+    	Tag tag = repositoryTag.findById(idTag);
+    	
+    	if(tag != null) {
+    		
+    		repositoryTag.delete(tag);
+    	
+    	}		
+    		
+    	return tag;
+    }
+    
+    @Override
+    public Tag updateTag(Tag tag) {
+    	
+    	tag = repositoryTag.save(tag);
+    	
+    	return tag;
+    	
+    }
+    
+    @Override
+    public Tag getTag(String keyUnity, Object valueUnity, String key, Object value) {
+    	return null;
+    }
+    
+    @Override
+    public List<UnityKnowledgeObject> getListTags(String keyUnity, Object valueUnity) {
+    	return null;
     }
     	
 }
